Kernel/Graphics: Migrate Intel driver to use the DisplayConnector design
